form setting update

+++ app/controllers/Admin.php
@@ -34,29 +34,32 @@ class Admin extends Controller
                 file_put_contents("uploads/index.php", "<?php //silence");
             }
 
-            
-            $allowed = ['image/jpeg', 'image/png'];
-            if (!empty($_FILES['image']['name'])) {
-                if ($_FILES['image']['error'] == 0) {
-                    if (in_array($_FILES['image']['type'], $allowed)) {
-                       $destination = $folder.time().$_FILES['image']['name'];
-                        move_uploaded_file($_FILES['image']['tmp_name'], $destination);
-                        $_POST['image'] = $destination; 
-                        if(file_exists($row->image)){
-                            unlink($row->image);
+            if ($user->edit_validate($data)) {
+
+                $allowed = ['image/jpeg', 'image/png'];
+                if (!empty($_FILES['image']['name'])) {
+                    if ($_FILES['image']['error'] == 0) {
+                        if (in_array($_FILES['image']['type'], $allowed)) {
+                            $destination = $folder . time() . $_FILES['image']['name'];
+                            move_uploaded_file($_FILES['image']['tmp_name'], $destination);
+                            $_POST['image'] = $destination;
+                            if (file_exists($row->image)) {
+                                unlink($row->image);
+                            }
+                        } else {
+                            $user->errors['image'] = "This file type not allowed";
                         }
                     } else {
-                        $user->errors['image'] = "This file type not allowed";
+                        $user->errors['image'] = "Could not uploada image";
                     }
-                } else {
-                    $user->errors['image'] = "Could not uploada image";
                 }
+                $user->update($id, $_POST);
+                redirect('admin/profile/' . $id);
             }
-            $user->update($id, $_POST);
-            redirect('admin/profile/' . $id);
         }
 
         $data['tittle'] = "Profile";
+        $data['errors'] = $user->errors;
         $this->view('admin/profile', $data);
     }
 }
